+++ server/src/main/java/org/jetbrains/bsp/bazel/server/sync/model/Library.kt
@@ -7,4 +7,5 @@ data class Library(
         val outputs: Set<URI>,
         val sources: Set<URI>,
         val dependencies: List<String>,
+        val interfaceJars: Set<URI> = emptySet(),
 )